+++ iridauploader/api/api_calls.py
@@ -487,6 +487,45 @@ class ApiCalls(object):
 
         return result
 
+    def get_metadata(self, sample_id, project_id):
+        """
+        API call to api/samples/{sampleId}/metadata
+        arguments:
+            sample_id
+            project_id
+        returns list of metadata associated with sampleID
+        """
+
+        logging.info("Getting metadata from sample ID '{}' found in project ID '{}'".format(sample_id, project_id))
+
+        try:
+            project_url = self._get_link(self.base_url, "projects")
+            sample_url = self._get_link(project_url, "project/samples",
+                                        target_dict={
+                                            "key": "identifier",
+                                            "value": project_id
+                                        })
+
+        except StopIteration:
+            logging.error("The given project ID doesn't exist: ".format(project_id))
+            raise exceptions.IridaResourceError("The given project ID doesn't exist", project_id)
+
+        try:
+            url = self._get_link(sample_url, "sample/metadata",
+                                 target_dict={
+                                     "key": "identifier",
+                                     "value": sample_id
+                                 })
+            response = self._session.get(url)
+
+        except StopIteration:
+            logging.error("The given sample doesn't exist: ".format(sample_id))
+            raise exceptions.IridaResourceError("The given sample ID doesn't exist", sample_id)
+
+        result = response.json()["resource"]["metadata"]
+
+        return result
+
     def send_project(self, project, clear_cache=True):
         """
         post request to send a project to IRIDA via API
@@ -627,6 +666,56 @@ class ApiCalls(object):
             json_res = json.loads(response.text)
         else:
             logging.error("Error while uploading [{}]: [{}]".format(sample_name, response.reason))
+            raise self._get_irida_exception(response)
+
+        return json_res
+
+    def send_metadata(self, metadata, project_id, sample_id):
+        """
+        Put request to add metadata to specific sample ID
+
+        :param metadata: Metadata object
+        :param project_id: id of project sample id is in
+        :param sample_id: id of sample to add metadata to
+        :return: json response from server
+        """
+
+        logging.info("Adding metadata to sample '{}' found in project '{}' on IRIDA.".format(sample_id, project_id))
+
+        try:
+            project_url = self._get_link(self.base_url, "projects")
+            sample_url = self._get_link(project_url, "project/samples",
+                                        target_dict={
+                                            "key": "identifier",
+                                            "value": project_id
+                                        })
+
+        except StopIteration:
+            logging.error("The given project ID doesn't exist: ".format(project_id))
+            raise exceptions.IridaResourceError("The given project ID doesn't exist", project_id)
+
+        try:
+            url = self._get_link(sample_url, "sample/metadata",
+                                 target_dict={
+                                     "key": "identifier",
+                                     "value": sample_id
+                                 })
+
+        except StopIteration:
+            logging.error("The given sample doesn't exist: ".format(sample_id))
+            raise exceptions.IridaResourceError("The given sample ID doesn't exist", sample_id)
+
+        json_obj = json.dumps(metadata.get_uploadable_dict())
+
+        headers_pkg = {'Content-Type': 'application/json'}
+
+        response = self._session.put(url, data=json_obj, headers=headers_pkg)
+
+        if response.status_code == 200:  # 200
+            json_res = json.loads(response.text)
+        else:
+            logging.error("Did not add metadata to sample. Response code is '{}' and error message is '{}'"
+                          "".format(response.status_code, response.text))
             raise self._get_irida_exception(response)
 
         return json_res