+++ CHANGELOG.md
@@ -10,6 +10,13 @@ Developer changes:
 * Changed `api` to post to `<base_url>/sequencingrun/<sequencing_run_type>` instead of `<base_url>/sequencingrun/miseqrun`
   * Added `sequencing_run_type` to the `parser -> model -> api` chain
 
+Beta 0.4.3
+----------
+Developer changes:
+* Added method `get_metadata(self, sample_id, project_id)` to allow other metadata to be fetched from the `api` module
+* Added method `send_metadata(self, metadata, project_id, sample_id)` to allow other metadata to be sent using the `api` module
+* Added model `Metadata` to support `send_metadata` method
+
 Beta 0.4.2
 ----------
 Added functionality: